+++ framework/helpers/BaseHtml.php
@@ -1359,7 +1359,23 @@ class BaseHtml
     public static function activeTextInput($model, $attribute, $options = [])
     {
         self::normalizeMaxLength($model, $attribute, $options);
+        self::activePlaceholder($model, $attribute, $options);
         return static::activeInput('text', $model, $attribute, $options);
+    }
+
+    /**
+     * Generate placeholder from model label.
+     * @param Model $model the model object
+     * @param string $attribute the attribute name or expression. See [[getAttributeName()]] for the format
+     * about attribute expression.
+     * @param array $options the tag options in terms of name-value pairs. These will be rendered as
+     * the attributes of the resulting tag. The values will be HTML-encoded using [[encode()]].
+     */
+    protected static function activePlaceholder($model, $attribute, &$options = [])
+    {
+        if (isset($options['placeholder']) && $options['placeholder'] === true) {
+            $options['placeholder'] = $model->getAttributeLabel($attribute);
+        }
     }
 
     /**
@@ -1458,6 +1474,7 @@ class BaseHtml
             $options['id'] = static::getInputId($model, $attribute);
         }
         self::normalizeMaxLength($model, $attribute, $options);
+        self::activePlaceholder($model, $attribute, $options);
         return static::textarea($name, $value, $options);
     }